Fix compile error after change to use C++23
list of compile error

/home/hpichu/work/dali-toolkit/dali-scene3d/internal/controls/model/model-impl.cpp:219:57: warning: bitwise operation between different enumeration types ‘Dali::CustomActorImpl::ActorFlags’ and ‘Dali::Toolkit::ControlImpl::ControlBehaviour’ is deprecated [-Wdeprecated-enum-enum-conversion]
  219 | : ControlImpl(ControlBehaviour(DISABLE_SIZE_NEGOTIATION | DISABLE_STYLE_CHANGE_SIGNALS)),

/home/hpichu/work/dali-toolkit/dali-scene3d/internal/loader/gltf2-util.cpp: In function ‘Dali::Scene3D::Loader::ModelRenderable* Dali::Scene3D::Loader::Internal::Gltf2Util::MakeModelRenderable(const gltf2::Mesh::Primitive&, Dali::Scene3D::Loader::Internal::Gltf2Util::ConversionContext&)’:
/home/hpichu/work/dali-toolkit/dali-scene3d/internal/loader/gltf2-util.cpp:1117:39: error: ‘constexpr gltf2::Named::Named()’ is protected within this context
 1117 |       ConvertMaterial(gltf2::Material{}, context.mOutput.mSceneMetadata.mImageMetadata, outMaterials, context);
      |                                       ^

/home/hpichu/work/dali-toolkit/automated-tests/src/dali-scene3d/utc-Dali-EnvironmentDefinition.cpp: In function ‘int UtcDaliEnvironmentDefinitionLoadRawFail()’:
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-scene3d/utc-Dali-EnvironmentDefinition.cpp:53:44: error: no matching function for call to ‘Dali::Scene3D::Loader::EnvironmentDefinition::EnvironmentDefinition(<brace-enclosed initializer list>)’
   53 |     EnvironmentDefinition envDef{name, name};

/home/hpichu/work/dali-toolkit/automated-tests/src/dali-scene3d-internal/utc-Dali-Gltf2LoaderImpl.cpp: In function ‘int UtcDaliGltfLoaderSuccess1()’:
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-scene3d-internal/utc-Dali-Gltf2LoaderImpl.cpp:380:3: error: could not convert ‘{<brace-enclosed initializer list>(), (((((((((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::ALBEDO) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::EMISSIVE)) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::OCCLUSION)) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::NORMAL)) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::SPECULAR)) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::SPECULAR_COLOR)) | ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::GLTF_CHANNELS)) | ((unsigned int)(128 << ((unsigned int)Dali::Scene3D::Loader::MaterialDefinition::ALPHA_CUTOFF_SHIFT)))), 0, Dali::Color::WHITE, 1.0e+0f, 0.0f, Dali::Vector4(((float)1.0e+0), ((float)7.6600000000000001e-1), ((float)3.3600000000000002e-1), ((float)1.0e+0)), 1.0e+0f, 1.0e+0f, Dali::Vector3(((float)2.0000000000000001e-1), ((float)1.0000000000000001e-1), ((float)0.0)), 1.0e+0f, 0.0f, 5.0e-1f, Dali::Vector3((float)0, (float)0, (float)1), true, false, false, false, true, false, Dali::Scene3D::Material::MASK, true, true, true, {{Dali::Scene3D::Loader::MaterialDefinition::ALBEDO, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}, {Dali::Scene3D::Loader::MaterialDefinition::NORMAL, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}, {Dali::Scene3D::Loader::MaterialDefinition::OCCLUSION, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}, {Dali::Scene3D::Loader::MaterialDefinition::EMISSIVE, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}, {Dali::Scene3D::Loader::MaterialDefinition::SPECULAR, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}, {Dali::Scene3D::Loader::MaterialDefinition::SPECULAR_COLOR, {"AnimatedCube_BaseColor.png", Dali::Scene3D::Loader::SamplerFlags::Encode(Dali::FilterMode::LINEAR_MIPMAP_LINEAR, Dali::FilterMode::LINEAR, Dali::WrapMode::CLAMP_TO_EDGE, Dali::WrapMode::REPEAT), Dali::Uint16Pair(256, 256), Dali::SamplingMode::BOX_THEN_NEAREST}}}, nullptr}’ from ‘<brace-enclosed initializer list>’ to ‘Dali::Scene3D::Loader::MaterialDefinition’
  380 |   };
      |   ^
      |   |
      |   <brace-enclosed initializer list>

/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h: In instantiation of ‘void TestDBusWrapper::Element::set(T&&, char) [with T = std::__cxx11::list<TestDBusWrapper::Element>]’:
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:488:10:   required from here
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:477:13: error: no matching function for call to ‘TestDBusWrapper::Element::get<std::__cxx11::list<TestDBusWrapper::Element, std::allocator<TestDBusWrapper::Element> > >()’
  477 |       get<T>()   = std::forward<T>(t);
      |       ~~~~~~^~
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:513:14: note: candidate: ‘template<class T, class> const T& TestDBusWrapper::Element::get() const’
  513 |     const T& get() const
      |              ^~~
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:513:14: note:   template argument deduction/substitution failed:
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:482:70: error: ISO C++ says that these are ambiguous, even though the worst conversion for the first is better than the worst conversion for the second: [-Werror]
  482 |     Element(T&& t, typename std::enable_if<detail::DBusSig<T>::value != 0>::type* = nullptr)
      |                                                                ~~~~~~^~~~
/home/hpichu/work/dali-toolkit/automated-tests/src/dali-toolkit-internal/../../../automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h:482:70: note: candidate 1: ‘operator!=(int, int)’ (built-in)

+++ automated-tests/src/dali-toolkit-internal/dali-toolkit-test-utils/dbus-wrapper.h
@@ -479,7 +479,7 @@ struct TestDBusWrapper : public DBusWrapper
 
   public:
     template<typename T>
-    Element(T&& t, typename std::enable_if<detail::DBusSig<T>::value != 0>::type* = nullptr)
+    Element(T&& t, typename std::enable_if<static_cast<int>(detail::DBusSig<T>::value) != 0>::type* = nullptr)
     {
       set(std::forward<T>(t));
     }
@@ -504,18 +504,18 @@ struct TestDBusWrapper : public DBusWrapper
     {
       return index_ == detail::IndexFromTypeTuple<ElementTuple, ElementList>::value;
     }
-    template<typename T, typename = typename std::enable_if<detail::DBusSig<T>::value != 0>::type>
+    template<typename T, typename = typename std::enable_if<static_cast<int>(detail::DBusSig<T>::value) != 0>::type>
     bool is() const
     {
       return index_ == detail::IndexFromTypeTuple<ElementTuple, T>::value;
     }
-    template<typename T, typename = typename std::enable_if<detail::DBusSig<T>::value != 0>::type>
+    template<typename T, typename = typename std::enable_if<static_cast<int>(detail::DBusSig<T>::value) != 0>::type>
     const T& get() const
     {
       if(!is<T>()) throw error{};
       return std::get<detail::IndexFromTypeTuple<ElementTuple, T>::value>(data);
     }
-    template<typename T, typename = typename std::enable_if<detail::DBusSig<T>::value != 0>::type>
+    template<typename T, typename = typename std::enable_if<static_cast<int>(detail::DBusSig<T>::value) != 0>::type>
     T& get()
     {
       if(!is<T>()) throw error{};
